feat: timestamp in description
